kernel/sched: Make sure the lowest priority is in 127 to 0
When preemption is enabled, we define K_LOWEST_THREAD_PRIO as
CONFIG_NUM_PREEMPT_PRIORITIES, which is limited from 0 to 128 by
kernel/Kconfig.  Unfortunately, using the number of priorities as the
thread priority is wrong.  The range of the priority for preemptible
threads must be 0 to 127.

K_LOWEST_THREAD_PRIO == 128 is currently used to create the idle
thread in init_idle_thread() when CONFIG_NUM_PREEMPT_PRIORITIES is
128.  The number 128 is passed down to z_setup_new_thread() ->
z_new_thread() -> z_new_thread_init() -> z_init_thread_base(), and
finally assigned to thread_base->prio which is type of s8_t.

Assigning 128 of type int to s8_t results in -127, which in turn means
now that the idle thread is the highest priority thread of all.  We
didn't notice this because we had special check for the idle thread in
should_preempt(), which is removed by the commit bd049626c5b04.

This patch only changes the definition of K_LOWEST_THREAD_PRIO but it
might be better to have some range check before the assignment to
thread_base->prio in z_init_thread_base() and z_thread_priority_set().

Signed-off-by: Yasushi SHOJI <y-shoji@ispace-inc.com>